requirements.txt
asgiref==3.2.10
boto3==1.14.51
botocore==1.17.51
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-import-export==2.3.0
django-storages==1.9.1
docutils==0.15.2
et-xmlfile==1.0.1
jdcal==1.4.1
jmespath==0.10.0
MarkupPy==1.14
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
python-barcode==0.13.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.10
xlrd==1.2.0
xlwt==1.3.0
